Fix dungeon units again, add seed and resolution to dungeon URL string
The problem with the internal grid being 10 foot squares is a continuing issue. This fix wall-papers over the problem, it does not fix it.
It will take a bigger rework to change that internal assumption without breaking the mapping, and I do not understand the mapping enough yet to undertake this.